--- frameioclient/services/__init__.py
@@ -1,7 +1,0 @@
-from .teams import Team
-from .users import User
-from .assets import Asset
-from .logs import AuditLogs
-from .comments import Comment
-from .projects import Project
-from .links import ReviewLink, PresentationLink
